Entities needed for calculation of MarginCall Implemented Entities needed for calculation of MarginCall Code split into sub directories CMakeLists added for sub directories

+++ Finance/CommonTypes/MoneyAmount.cpp
@@ -28,7 +28,7 @@ MoneyAmount* MoneyAmount::convertFromDollar(std::string toCurrencyName){
 }
 
 
-float MoneyAmount::getQuantity(){
+double MoneyAmount::getQuantity(){
 	return quantity;
 }
 Currency* MoneyAmount::getCurrency(){
